fixed: recommendedProjects
